requirements.txt
python-telegram-bot==22.6
motor==3.6.0
pymongo>=4.9,<4.10
python-dotenv>=1.0.0
